requirements.txt
mss
pynput
pillow
